Refactor selector view into modular package and add tests
