+++ src/prompt_automation/gui/selector/view.py
@@ -1,872 +1,10 @@
-"""Tkinter view and callbacks for the template selector GUI.
+from importlib import import_module
+import os
 
-Features:
-- Hierarchical browsing (directories & templates) starting at PROMPTS_DIR
-- Inline filtering (current directory) AND recursive search (toggle 'Recursive')
-- Multi-select with synthetic combined template output (Finish Multi)
-- Content-aware search: matches path, title, placeholder names, template body
-- Preview window for template contents
-- Override management dialogs
-"""
+# Allow access to submodules in the accompanying ``view`` package directory
+__path__ = [os.path.join(os.path.dirname(__file__), "view")]
 
-from __future__ import annotations
-import tkinter as tk
-from tkinter import ttk, messagebox
-from pathlib import Path
-from typing import List, Optional
+# Re-export the main view class
+SelectorView = import_module(__name__ + ".orchestrator").SelectorView
 
-from .model import ListingItem, TemplateEntry, BrowserState
-from ..new_template_wizard import open_new_template_wizard
-from ...errorlog import get_logger
-from ...starred import load_starred, toggle_star, MAX_STARRED
-
-_log = get_logger(__name__)
-
-# --- Preview window -------------------------------------------------------
-
-def _open_preview(parent: tk.Tk, entry: TemplateEntry) -> None:
-    try:
-        tmpl = entry.data
-        preview = tk.Toplevel(parent)
-        preview.title(f"Preview: {tmpl.get('title', entry.path.name)}")
-        preview.geometry("700x500")
-        preview.resizable(True, True)
-        from ..fonts import get_display_font
-        text = tk.Text(preview, wrap="word", font=get_display_font(master=parent))
-        text.pack(fill="both", expand=True)
-        lines = tmpl.get('template', [])
-        text.insert("1.0", "\n".join(lines))
-        text.config(state="disabled")
-        preview.transient(parent)
-        preview.grab_set()
-    except Exception as e:
-        messagebox.showerror("Preview Error", str(e))
-
-# --- Manage overrides dialog ----------------------------------------------
-
-def _manage_overrides(root: tk.Tk, service):
-    """Unified manager for file & simple value overrides with inline editing."""
-    win = tk.Toplevel(root)
-    win.title("Manage Overrides")
-    win.geometry("760x420")
-    frame = tk.Frame(win, padx=12, pady=12)
-    frame.pack(fill="both", expand=True)
-    hint = tk.Label(
-        frame,
-        text="Double‑click a row to edit value (simple overrides). Delete removes. File overrides show path/skip.",
-        wraplength=720,
-        justify="left",
-        fg="#555",
-    )
-    hint.pack(anchor="w", pady=(0, 6))
-    cols = ("kind","tid","name","data")
-    tree = ttk.Treeview(frame, columns=cols, show="headings")
-    widths = {"kind":80, "tid":60, "name":160, "data":360}
-    for c in cols:
-        tree.heading(c, text=c.capitalize())
-        tree.column(c, width=widths[c], anchor="w")
-    sb = tk.Scrollbar(frame, orient="vertical", command=tree.yview)
-    tree.configure(yscrollcommand=sb.set)
-    tree.pack(side="left", fill="both", expand=True)
-    sb.pack(side="right", fill="y")
-
-    import json
-    def _refresh():
-        tree.delete(*tree.get_children())
-        for tid, name, info in service.list_file_overrides():
-            tree.insert("", "end", values=("file", tid, name, json.dumps(info)))
-        for tid, name, val in service.list_template_value_overrides():
-            # show scalar/list succinctly
-            if isinstance(val, list):
-                display = ", ".join(str(v) for v in val[:5]) + (" …" if len(val) > 5 else "")
-            else:
-                display = str(val)
-            tree.insert("", "end", values=("value", tid, name, display))
-    _refresh()
-
-    def _edit(event=None):
-        sel = tree.selection()
-        if not sel: return
-        item = tree.item(sel[0]); kind, tid, name, data = item['values']
-        if kind != 'value':
-            return  # only simple values editable
-        dlg = tk.Toplevel(win)
-        dlg.title(f"Edit Override: {tid}/{name}")
-        tk.Label(dlg, text=f"Template {tid} – {name}").pack(padx=10,pady=(10,4))
-        txt = tk.Text(dlg, width=60, height=6, wrap='word')
-        txt.pack(padx=10, pady=4)
-        txt.insert('1.0', data)
-        def _ok():
-            val = txt.get('1.0','end-1c').strip()
-            service.set_template_value_override(int(tid), name, val)
-            _refresh(); dlg.destroy()
-        tk.Button(dlg, text='Save', command=_ok).pack(side='left', padx=10, pady=8)
-        tk.Button(dlg, text='Cancel', command=dlg.destroy).pack(side='left', padx=4, pady=8)
-        dlg.transient(win); dlg.grab_set(); txt.focus_set()
-        dlg.bind('<Escape>', lambda e: (dlg.destroy(),'break'))
-        dlg.bind('<Return>', lambda e: (_ok(),'break'))
-    tree.bind('<Double-1>', _edit)
-
-    btns = tk.Frame(win); btns.pack(pady=8)
-    def do_remove():
-        sel = tree.selection()
-        if not sel: return
-        item = tree.item(sel[0]); kind, tid, name, _ = item['values']
-        removed = False
-        if kind == 'file':
-            removed = service.reset_single_file_override(int(tid), name)
-        else:
-            removed = service.reset_template_value_override(int(tid), name)
-        if removed:
-            tree.delete(sel[0])
-    tk.Button(btns, text="Remove Selected", command=do_remove).pack(side="left", padx=4)
-    tk.Button(btns, text="Close", command=win.destroy).pack(side="left", padx=4)
-
-# --- Main selector --------------------------------------------------------
-
-def _edit_exclusions(root: tk.Tk, service):  # extracted for shared menu use
-    try:
-        import json
-        _PD = service.PROMPTS_DIR
-    except Exception:
-        return
-    dlg = tk.Toplevel(root)
-    dlg.title("Edit Global Exclusions (exclude_globals)")
-    dlg.geometry("640x400")
-    tk.Label(dlg, text="Enter template ID (numeric) or browse to load its metadata.").pack(anchor='w', padx=10, pady=(10,4))
-    topf = tk.Frame(dlg); topf.pack(fill='x', padx=10)
-    id_var = tk.StringVar()
-    tk.Entry(topf, textvariable=id_var, width=10).pack(side='left')
-    status_var = tk.StringVar(value="")
-    tk.Label(dlg, textvariable=status_var, fg="#555").pack(anchor='w', padx=10, pady=(4,4))
-    txt = tk.Text(dlg, wrap='word')
-    txt.pack(fill='both', expand=True, padx=10, pady=6)
-    txt.insert('1.0', "# Enter one global key per line to exclude for this template\n")
-    current_path: list[Path] = []
-    def _load():
-        tid = id_var.get().strip()
-        if not tid.isdigit():
-            status_var.set("Template id must be numeric")
-            return
-        target = None
-        for p in _PD.rglob("*.json"):
-            try:
-                data = json.loads(p.read_text())
-            except Exception:
-                continue
-            if data.get('id') == int(tid):
-                target = (p, data); break
-        if not target:
-            status_var.set("Template not found")
-            return
-        p, data = target
-        current_path.clear(); current_path.append(p)
-        meta = data.get('metadata') if isinstance(data.get('metadata'), dict) else {}
-        raw_ex = meta.get('exclude_globals')
-        lines = []
-        if isinstance(raw_ex, (list, tuple)):
-            lines = [str(x) for x in raw_ex]
-        elif isinstance(raw_ex, str):
-            if ',' in raw_ex:
-                lines = [s.strip() for s in raw_ex.split(',') if s.strip()]
-            elif raw_ex.strip():
-                lines = [raw_ex.strip()]
-        txt.delete('1.0','end')
-        if lines:
-            txt.insert('1.0', "\n".join(lines))
-        status_var.set(f"Loaded {p.name}")
-    def _save():
-        if not current_path:
-            status_var.set("Load a template first")
-            return
-        p = current_path[0]
-        try:
-            data = json.loads(p.read_text())
-        except Exception as e:
-            status_var.set(f"Read error: {e}")
-            return
-        meta = data.get('metadata') if isinstance(data.get('metadata'), dict) else {}
-        if not isinstance(meta, dict):
-            meta = {}; data['metadata'] = meta
-        raw = [l.strip() for l in txt.get('1.0','end-1c').splitlines() if l.strip() and not l.strip().startswith('#')]
-        if raw:
-            meta['exclude_globals'] = raw
-        else:
-            meta.pop('exclude_globals', None)
-        try:
-            p.write_text(json.dumps(data, indent=2))
-            status_var.set("Saved")
-        except Exception as e:
-            status_var.set(f"Write error: {e}")
-    tk.Button(topf, text="Load", command=_load).pack(side='left', padx=6)
-    tk.Button(topf, text="Save", command=_save).pack(side='left')
-    tk.Button(topf, text="Close", command=dlg.destroy).pack(side='right')
-    dlg.transient(root); dlg.grab_set(); dlg.focus_set()
-
-def open_template_selector(service, embedded: bool = False, parent=None) -> Optional[dict]:
-    """Open the template selector.
-
-    When embedded=True, reuse the provided parent (frame/root) and avoid creating
-    a new Tk root so that OS window snapping is preserved.
-    """
-    if embedded and parent is None:
-        embedded = False
-
-    if embedded:
-        # In embedded mode we still want a real toplevel for menus/bindings.
-        host = parent.winfo_toplevel()  # type: ignore[attr-defined]
-        root = host  # treat host toplevel as root for bindings/menus
-        container = tk.Frame(parent)
-        container.pack(fill="both", expand=True)
-    else:
-        root = tk.Tk()
-        root.title("Select Template - Prompt Automation")
-        root.geometry("950x720")
-        root.resizable(True, True)
-        root.lift(); root.focus_force(); root.attributes("-topmost", True); root.after(120, lambda: root.attributes("-topmost", False))
-        container = root
-
-
-    # Build two browsers: one for shared (PROMPTS_DIR), and optionally one for local/private
-    browser = service.create_browser_state()
-    browser.build()
-    # Try to build local/private browser if exists
-    _PD = service.PROMPTS_DIR
-    local_dir = _PD.parent / "local"
-    local_browser: Optional[BrowserState] = None
-    try:
-        if local_dir.exists() and local_dir.is_dir():
-            local_browser = BrowserState(local_dir)
-            local_browser.build()
-    except Exception:
-        local_browser = None
-
-    # Menu replaced with shared options menu
-    from ..options_menu import configure_options_menu
-    # Pass this module object directly so options menu can access internal callbacks
-    import sys as _sys
-    _this_module = _sys.modules.get(__name__)
-    accels = configure_options_menu(
-        root,
-        selector_view_module=_this_module,
-        selector_service=service,
-        include_global_reference=False,
-        include_manage_templates=False,
-    )
-    for seq, fn in accels.items():
-        root.bind(seq, lambda e, f=fn: (f(), 'break'))
-
-    # Top controls: search box & multi-select toggle
-    top = tk.Frame(container, padx=10, pady=6); top.pack(fill="x")
-    tk.Label(top, text="Search:").pack(side="left")
-    search_var = tk.StringVar()
-    search_entry = tk.Entry(top, textvariable=search_var, width=42)
-    search_entry.pack(side="left", padx=(4,10))
-    # Default to recursive search; user can opt-out (non-recursive)
-    non_recursive_var = tk.BooleanVar(value=False)
-    tk.Checkbutton(top, text="Non-recursive", variable=non_recursive_var).pack(side="left", padx=(0,8))
-    multi_var = tk.BooleanVar(value=False)
-    tk.Checkbutton(top, text="Multi-select", variable=multi_var).pack(side="left")
-    preview_btn = tk.Button(top, text="Preview", state="disabled")
-    preview_btn.pack(side="left", padx=6)
-    breadcrumb_var = tk.StringVar(value=browser.breadcrumb())
-    breadcrumb_lbl = tk.Label(top, textvariable=breadcrumb_var, anchor="w", fg="#555", wraplength=600)
-    breadcrumb_lbl.pack(side="left", fill="x", expand=True, padx=(12,0))
-
-    # Listbox
-    main_frame = tk.Frame(container, padx=10, pady=6)
-    main_frame.pack(fill="both", expand=True)
-    listbox = tk.Listbox(main_frame, font=("Arial", 10), selectmode="extended")
-    sb = tk.Scrollbar(main_frame, orient="vertical", command=listbox.yview)
-    listbox.config(yscrollcommand=sb.set)
-    listbox.pack(side="left", fill="both", expand=True)
-    sb.pack(side="right", fill="y")
-
-    selected_template: Optional[dict] = None
-    preview_win: Optional[tk.Toplevel] = None  # for toggle behaviour (Ctrl+P)
-    multi_selected: List[dict] = []
-
-    # Track which domain is active ('shared' or 'local')
-    active_domain = 'shared'
-    # flat_items holds ListingItem objects (may include headers/back) for display
-    flat_items: List[ListingItem] = []
-
-    def refresh(display_items: Optional[List[ListingItem]] = None):
-        """Rebuild listbox based on active domain.
-
-        Shared domain: show current shared directory. If at shared root and local exists,
-        append a header then the top-level local styles (only root of local domain).
-        Local domain: show only local browser items. Provide a '[..] Shared Root' back item
-        when at local root to return to shared domain.
-        """
-        nonlocal selected_template, flat_items, active_domain
-        listbox.delete(0, 'end')
-        flat_items = []
-        show_extra_sections = (
-            active_domain == 'shared'
-            and browser.current == browser.root
-            and not search_var.get().strip()
-            and not non_recursive_var.get()  # only at top-level unfiltered
-        )
-        # Starred section
-        if show_extra_sections:
-            starred = load_starred()
-            # Filter out missing templates silently
-            valid_starred: list[tuple[str, dict]] = []  # (rel, tmpl)
-            for rel in starred:
-                data = None
-                if rel.startswith('local/') and local_browser:
-                    # local relative path: local/<relative inside local root>
-                    sub = rel.split('/', 1)[1] if '/' in rel else ''
-                    p = (local_browser.root / sub) if sub else None
-                    if p and p.exists():
-                        try:
-                            from ...renderer import load_template
-                            data = load_template(p)
-                        except Exception:
-                            data = None
-                else:
-                    data = service.load_template_by_relative(rel)
-                if data and 'template' in data:
-                    valid_starred.append((rel, data))
-            if valid_starred:
-                flat_items.append(ListingItem(type='header', display='--- Starred ---'))
-                for rel, data in valid_starred:
-                    li = ListingItem(type='template', path=None, template=TemplateEntry(path=None, data=data), display=f"{data.get('title', rel)}")
-                    setattr(li, 'star_rel', rel)
-                    setattr(li, 'is_starred_section', True)
-                    flat_items.append(li)
-        # Quick command section
-        if show_extra_sections:
-            mapping = service.load_shortcuts()
-            # Only digits 0-9
-            display_pairs = []
-            for d in '0123456789':
-                rel = mapping.get(d)
-                if not rel:
-                    continue
-                data = service.load_template_by_relative(rel)
-                if data and 'template' in data:
-                    display_pairs.append((d, rel, data))
-            if display_pairs:
-                flat_items.append(ListingItem(type='header', display='--- Quick Commands ---'))
-                for digit, rel, data in display_pairs:
-                    li = ListingItem(type='template', path=None, template=TemplateEntry(path=None, data=data), display=f"{digit}: {data.get('title', rel)}")
-                    setattr(li, 'quick_digit', digit)
-                    setattr(li, 'quick_rel', rel)
-                    setattr(li, 'is_quick_section', True)
-                    flat_items.append(li)
-        # Normal listing afterwards
-        inserted_any_section = show_extra_sections and any(
-            it.type == 'header' and it.display.strip() in {'--- Starred ---','--- Quick Commands ---'} for it in flat_items
-        )
-        if active_domain == 'shared':
-            if inserted_any_section:
-                # Add MAIN header spacer before core directory/template listing
-                flat_items.append(ListingItem(type='header', display='--- MAIN ---'))
-            base_items = display_items if display_items is not None else browser.items
-            if base_items:
-                flat_items.extend(base_items)
-            if local_browser and browser.current == browser.root:
-                header = ListingItem(type='header', display='--- Local (Private) * ---')
-                flat_items.append(header)
-                for it in local_browser.items:
-                    if it.type == 'up':
-                        continue
-                    marker = ListingItem(type=it.type, path=it.path, template=it.template,
-                                         display=(f"* {it.display}" if it.type == 'template' else it.display))
-                    setattr(marker, 'origin', 'local')
-                    flat_items.append(marker)
-        else:  # local domain
-            if local_browser:
-                if local_browser.current == local_browser.root:
-                    back = ListingItem(type='back', display='[..] Shared Root')
-                    flat_items.append(back)
-                for it in local_browser.items:
-                    if local_browser.current == local_browser.root and it.type == 'up':
-                        continue
-                    marker = ListingItem(type=it.type, path=it.path, template=it.template,
-                                         display=(f"* {it.display}" if it.type == 'template' else it.display))
-                    setattr(marker, 'origin', 'local')
-                    flat_items.append(marker)
-        if not flat_items:
-            flat_items.append(ListingItem(type='empty', display='<empty>'))
-        for it in flat_items:
-            listbox.insert('end', it.display)
-        for idx, it in enumerate(flat_items):
-            if it.type not in {'header','empty'}:
-                listbox.selection_set(idx); listbox.see(idx); break
-        selected_template = None
-        preview_btn.config(state='disabled')
-        if active_domain == 'shared':
-            breadcrumb_var.set(browser.breadcrumb())
-        else:
-            if local_browser:
-                if local_browser.current == local_browser.root:
-                    breadcrumb_var.set('Local')
-                else:
-                    breadcrumb_var.set(f"Local/{local_browser.current.relative_to(local_browser.root)}")
-            else:
-                breadcrumb_var.set('Local')
-
-    refresh()
-
-    # Ensure initial keyboard focus lands in the search box for immediate typing
-    # (Some WMs ignore early focus requests; schedule a couple of delayed attempts.)
-    def _focus_initial():
-        try:
-            search_entry.focus_set()
-            search_entry.focus_force()
-            search_entry.select_range(0, 'end')
-        except Exception:
-            pass
-    # Immediate + delayed to cope with window managers / animation
-    _focus_initial()
-    root.after(80, _focus_initial)
-    root.after(200, _focus_initial)
-
-    def current_items() -> List[ListingItem]:
-        """Return items for the active domain (without any headers/back)."""
-        q = search_var.get().strip()
-        if active_domain == 'local' and local_browser:
-            if not q:
-                return local_browser.items
-            if non_recursive_var.get():
-                return local_browser.filter(q)
-            return local_browser.search(q)
-        # Shared domain
-        if not q:
-            return browser.items
-        if non_recursive_var.get():
-            return browser.filter(q)
-        return browser.search(q)
-
-    def on_search(*_):
-        refresh(current_items())
-    search_var.trace_add("write", on_search)
-    non_recursive_var.trace_add("write", on_search)
-
-    # Quick key: 's' focuses search box and toggles recursion mode
-    def focus_search(event=None):
-        # Just focus and select text; do not toggle recursion mode
-        search_entry.focus_set()
-        search_entry.after(10, lambda: search_entry.select_range(0, 'end'))
-        return "break"
-    root.bind("s", focus_search)
-    listbox.bind("s", focus_search)
-
-    def get_selected_item() -> Optional[ListingItem]:
-        sel = listbox.curselection()
-        if not sel:
-            return None
-        idx = sel[0]
-        if idx < 0 or idx >= len(flat_items):
-            return None
-        it = flat_items[idx]
-        if it.type in {'header','empty'}:
-            return None
-        return it
-
-    def open_or_select():
-        nonlocal selected_template, multi_selected, active_domain
-        item = get_selected_item()
-        if not item:
-            return
-        # Starred or quick sections behave like normal templates
-        if getattr(item, 'is_starred_section', False) or getattr(item, 'is_quick_section', False):
-            if item.template:
-                selected_template = item.template.data
-                if embedded:
-                    try: container.destroy()
-                    except Exception: pass
-                else:
-                    root.destroy()
-            return
-        origin = getattr(item, 'origin', active_domain)
-        # Back navigation
-        if item.type == 'back':
-            active_domain = 'shared'
-            refresh(current_items()); return
-        if item.type in {'up','dir'}:
-            if origin == 'local' and local_browser:
-                if active_domain != 'local':
-                    active_domain = 'local'
-                if item.type == 'up':
-                    for it2 in local_browser.items:
-                        if it2.type == 'up':
-                            local_browser.enter(it2); break
-                else:
-                    for it2 in local_browser.items:
-                        if it2.path == item.path:
-                            local_browser.enter(it2); break
-                refresh(current_items()); return
-            else:
-                active_domain = 'shared'
-                if item.type == 'up':
-                    for it2 in browser.items:
-                        if it2.type == 'up':
-                            browser.enter(it2); break
-                else:
-                    for it2 in browser.items:
-                        if it2.path == item.path:
-                            browser.enter(it2); break
-                refresh(current_items()); return
-        if item.type == 'template' and item.template:
-            if multi_var.get():
-                tmpl_dict = item.template.data
-                if tmpl_dict in multi_selected:
-                    multi_selected.remove(tmpl_dict)
-                else:
-                    multi_selected.append(tmpl_dict)
-                idx = listbox.curselection()[0]
-                disp = item.display
-                if disp.startswith('* '):
-                    disp = disp[2:]
-                else:
-                    disp = '* ' + disp
-                listbox.delete(idx); listbox.insert(idx, disp); listbox.selection_set(idx)
-            else:
-                selected_template = item.template.data
-                if embedded:
-                    try: container.destroy()
-                    except Exception: pass
-                else:
-                    root.destroy()
-
-    # --- Shortcut key handling (numeric keys mapped to specific templates) ---
-    def _on_digit(event):
-        if event.char and event.char.isdigit():
-            tmpl = service.resolve_shortcut(event.char)
-            if tmpl:
-                nonlocal selected_template
-                selected_template = tmpl
-                if embedded:
-                    try: container.destroy()
-                    except Exception: pass
-                else:
-                    root.destroy()
-                return "break"
-        return None
-    # Bind 0-9 keys globally (without modifiers) when focus is in list or search
-    for d in "0123456789":
-        root.bind(d, _on_digit)
-        listbox.bind(d, _on_digit)
-        search_entry.bind(d, _on_digit)
-
-    def on_preview():
-        item = get_selected_item()
-        if item and item.type == 'template' and item.template:
-            _open_preview(root, item.template)
-    preview_btn.config(command=on_preview)
-
-    def on_select_event(event=None):
-        open_or_select(); return "break"
-
-    listbox.bind("<Return>", on_select_event)
-    listbox.bind("<KP_Enter>", on_select_event)
-    listbox.bind("<Double-Button-1>", on_select_event)
-
-    def on_backspace(event):
-        nonlocal active_domain
-        if active_domain == 'local' and local_browser:
-            if local_browser.current != local_browser.root:
-                for it in local_browser.items:
-                    if it.type == 'up':
-                        local_browser.enter(it); refresh(current_items()); break
-                return 'break'
-            else:
-                active_domain = 'shared'; refresh(current_items()); return 'break'
-        if browser.current != browser.root:
-            for it in browser.items:
-                if it.type == 'up':
-                    browser.enter(it); refresh(current_items()); break
-            return 'break'
-        return None
-    listbox.bind("<BackSpace>", on_backspace)
-
-    def on_key_release(event):
-        item = get_selected_item()
-        if item and item.type=='template':
-            preview_btn.config(state="normal")
-        else:
-            preview_btn.config(state="disabled")
-    listbox.bind("<<ListboxSelect>>", on_key_release)
-    listbox.bind("<KeyRelease-Up>", on_key_release)
-    listbox.bind("<KeyRelease-Down>", on_key_release)
-
-    # --- Keyboard interaction while focus is in the search box -------------
-    def _move_selection(delta: int):
-        size = len(flat_items)
-        if size == 0:
-            return
-        sel = listbox.curselection()
-        if not sel:
-            idx = 0 if delta >= 0 else size - 1
-        else:
-            idx = sel[0] + delta
-        idx = max(0, min(size - 1, idx))
-        # Skip headers
-        step = 1 if delta >= 0 else -1
-        while 0 <= idx < size and flat_items[idx].type == 'header':
-            idx += step
-        idx = max(0, min(size - 1, idx))
-        listbox.selection_clear(0, 'end')
-        listbox.selection_set(idx)
-        listbox.see(idx)
-        # Trigger preview button state update
-        on_key_release(None)
-
-    def _on_search_nav_down(event):
-        _move_selection(1)
-        return "break"
-
-    def _on_search_nav_up(event):
-        _move_selection(-1)
-        return "break"
-
-    def _on_search_enter(event):
-        # Use current highlighted item
-        open_or_select()
-        return "break"
-
-    search_entry.bind('<Down>', _on_search_nav_down)
-    search_entry.bind('<Up>', _on_search_nav_up)
-    search_entry.bind('<Return>', _on_search_enter)
-    search_entry.bind('<KP_Enter>', _on_search_enter)
-
-    # Preview toggle (Ctrl+P)
-    def toggle_preview(event=None):
-        nonlocal preview_win
-        item = get_selected_item()
-        if not item or item.type != 'template' or not item.template:
-            return "break"
-        # Close existing preview if open
-        if preview_win and preview_win.winfo_exists():
-            preview_win.destroy()
-            preview_win = None
-            return "break"
-        # Open new preview window and keep reference
-        try:
-            tmpl = item.template.data
-            from ..fonts import get_display_font
-            preview_win = tk.Toplevel(root)
-            preview_win.title(f"Preview: {tmpl.get('title', item.path.name)}")
-            preview_win.geometry("700x500")
-            preview_win.resizable(True, True)
-            txt = tk.Text(preview_win, wrap='word', font=get_display_font(master=preview_win))
-            txt.pack(fill='both', expand=True)
-            lines = tmpl.get('template', [])
-            txt.insert('1.0', "\n".join(lines))
-            txt.config(state='disabled')
-            def _closed(_):
-                nonlocal preview_win
-                preview_win = None
-            preview_win.bind('<Destroy>', _closed)
-        except Exception as e:  # pragma: no cover
-            messagebox.showerror("Preview Error", str(e))
-        return "break"
-
-    root.bind('<Control-p>', toggle_preview)
-    listbox.bind('<Control-p>', toggle_preview)
-    search_entry.bind('<Control-p>', toggle_preview)
-
-    # Toggle recursive search mode (Ctrl+L) for convenience
-    def _toggle_recursive(event=None):
-        non_recursive_var.set(not non_recursive_var.get())
-        on_search(); return 'break'
-    root.bind('<Control-l>', _toggle_recursive)
-    search_entry.bind('<Control-l>', _toggle_recursive)
-
-    # Star toggle (logic defined before buttons so button & keybinding share it)
-    def _toggle_star(event=None):  # pragma: no cover - GUI only
-        item = get_selected_item()
-        if not item or not item.template:
-            return 'break'
-        # Determine rel path
-        rel: Optional[str] = getattr(item, 'star_rel', None)
-        if not rel:
-            # normal shared template
-            if getattr(item, 'origin', 'shared') == 'shared' and item.path:
-                try:
-                    rel = str(item.path.relative_to(service.PROMPTS_DIR))
-                except Exception:
-                    rel = None
-            elif getattr(item, 'origin', None) == 'local' and item.path and local_browser:
-                try:
-                    rel = f"local/{item.path.relative_to(local_browser.root)}"
-                except Exception:
-                    rel = None
-            # quick command section item
-            if not rel and getattr(item, 'quick_rel', None):
-                rel = getattr(item, 'quick_rel')
-        if not rel:
-            return 'break'
-        current = load_starred()
-        if rel in current:
-            toggle_star(rel)
-            refresh(current_items())
-        else:
-            if len(current) >= MAX_STARRED:
-                try:
-                    messagebox.showerror('Starred Limit', f'Max {MAX_STARRED} starred templates reached. Unstar another first.')
-                except Exception:
-                    pass
-                return 'break'
-            toggle_star(rel)
-            refresh(current_items())
-        return 'break'
-
-    root.bind('*', _toggle_star)
-    listbox.bind('*', _toggle_star)
-    search_entry.bind('*', _toggle_star)
-
-    # Buttons
-    btns = tk.Frame(container, pady=6); btns.pack(fill="x")
-    def finish_multi():
-        nonlocal selected_template
-        if multi_selected:
-            # wrap multiple templates into a synthetic combined one (concatenate template arrays)
-            combined = {
-                'id': -1,
-                'title': f"Multi ({len(multi_selected)})",
-                'style': 'multi',
-                'template': sum([t.get('template', []) for t in multi_selected], []),
-                'placeholders': [],
-            }
-            selected_template = combined
-            if embedded:
-                try:
-                    container.destroy()
-                except Exception:
-                    pass
-            else:
-                root.destroy()
-    tk.Button(btns, text="Open / Select (Enter)", command=open_or_select, padx=18).pack(side="left", padx=(0,8))
-    tk.Button(btns, text="Finish Multi", command=finish_multi).pack(side="left", padx=(0,8))
-    tk.Button(btns, text="Preview", command=on_preview).pack(side="left", padx=(0,8))
-    tk.Button(btns, text="Star / Unstar (*)", command=_toggle_star).pack(side="left", padx=(0,8))
-    tk.Button(btns, text="Cancel (Esc)", command=root.destroy).pack(side="left")
-
-    def _finish():
-        if embedded:
-            try:
-                container.destroy()
-            except Exception:
-                pass
-        else:
-            root.destroy()
-    root.bind("<Escape>", lambda e: (_finish(), "break"))
-
-    if embedded:
-        # Poll until selection set; run nested loop to remain responsive
-        while container.winfo_exists() and selected_template is None:
-            try:
-                container.update(); container.update_idletasks()
-            except Exception:
-                break
-        return selected_template
-    else:
-        root.mainloop()
-        return selected_template
-
-# --- Shortcuts manager dialog ----------------------------------------------
-
-def _manage_shortcuts(root: tk.Tk, service):  # pragma: no cover - GUI only
-    win = tk.Toplevel(root)
-    win.title("Manage Template Shortcuts")
-    win.geometry("700x420")
-    win.resizable(True, True)
-    frame = tk.Frame(win, padx=12, pady=12); frame.pack(fill="both", expand=True)
-    hint = tk.Label(frame, text="Assign single-digit keys to frequently used templates. Double-click a row to set/edit key. Use Renumber to rename files so IDs match chosen digits.", wraplength=660, justify="left", fg="#444")
-    hint.pack(anchor="w", pady=(0,8))
-    cols = ("key","rel","title","id")
-    tree = ttk.Treeview(frame, columns=cols, show="headings")
-    widths = {"key":60, "rel":300, "title":200, "id":60}
-    for c in cols:
-        tree.heading(c, text=c.upper())
-        tree.column(c, width=widths[c], anchor="w")
-    vsb = tk.Scrollbar(frame, orient="vertical", command=tree.yview)
-    tree.configure(yscrollcommand=vsb.set)
-    tree.pack(side="left", fill="both", expand=True); vsb.pack(side="right", fill="y")
-
-    mapping = service.load_shortcuts()  # key -> rel path
-
-    # Gather templates for listing
-    _PD = service.PROMPTS_DIR
-    templates: list[tuple[str, Path, dict]] = []  # (rel, path, data)
-    for p in sorted(_PD.rglob("*.json")):
-        rel = str(p.relative_to(_PD))
-        data = service.load_template_by_relative(rel)
-        if data and "template" in data:
-            templates.append((rel, p, data))
-
-    def _refresh_tree():
-        tree.delete(*tree.get_children())
-        for rel, path, data in templates:
-            key = next((k for k,v in mapping.items() if v == rel), "")
-            tree.insert("", "end", values=(key, rel, data.get("title",""), data.get("id","")))
-    _refresh_tree()
-
-    def _edit_key(event=None):
-        sel = tree.selection()
-        if not sel: return
-        item = tree.item(sel[0])
-        cur_key, rel, *_ = item["values"]
-        def _apply(new_key: str):
-            new_key = new_key.strip()
-            # Remove existing mapping for this rel or key
-            for k in list(mapping.keys()):
-                if mapping[k] == rel or k == new_key:
-                    mapping.pop(k, None)
-            if new_key:
-                mapping[new_key] = rel
-            service.save_shortcuts(mapping)
-            _refresh_tree()
-        # Simple prompt dialog
-        dlg = tk.Toplevel(win); dlg.title("Set Key")
-        tk.Label(dlg, text=f"Template: {rel}\nCurrent key: {cur_key or '(none)'}\nEnter new single-digit key (blank to clear):").pack(padx=12,pady=12)
-        var = tk.StringVar(value=cur_key)
-        ent = tk.Entry(dlg, textvariable=var); ent.pack(padx=12, pady=(0,12))
-        def _ok():
-            val = var.get()
-            if val and (len(val)!=1 or not val.isdigit()):
-                messagebox.showerror("Validation", "Key must be a single digit 0-9 or blank")
-                return
-            _apply(val)
-            dlg.destroy()
-        tk.Button(dlg, text="OK", command=_ok).pack(side="left", padx=12, pady=8)
-        tk.Button(dlg, text="Cancel", command=dlg.destroy).pack(side="left", padx=4, pady=8)
-        ent.focus_set(); ent.select_range(0,'end')
-        dlg.bind('<Return>', lambda e: (_ok(), 'break'))
-        dlg.bind('<Escape>', lambda e: (dlg.destroy(), 'break'))
-        dlg.transient(win); dlg.grab_set()
-
-    tree.bind('<Double-1>', _edit_key)
-
-    btns = tk.Frame(frame); btns.pack(fill="x", pady=8)
-    def _do_delete():
-        sel = tree.selection();
-        if not sel: return
-        item = tree.item(sel[0]); cur_key, rel, *_ = item['values']
-        # remove mapping
-        for k in list(mapping.keys()):
-            if mapping[k] == rel:
-                mapping.pop(k, None)
-        service.save_shortcuts(mapping); _refresh_tree()
-    def _do_renumber():
-        # Apply renumbering according to numeric shortcut keys
-        try:
-            service.renumber_templates(mapping)
-            messagebox.showinfo("Renumber", "Renumbering complete. Close and reopen selector to see changes.")
-        except Exception as e:
-            messagebox.showerror("Renumber", f"Failed: {e}")
-    tk.Button(btns, text="Edit Key", command=_edit_key).pack(side="left", padx=4)
-    tk.Button(btns, text="Clear Key", command=_do_delete).pack(side="left", padx=4)
-    tk.Button(btns, text="Renumber", command=_do_renumber).pack(side="left", padx=10)
-    tk.Button(btns, text="Close", command=win.destroy).pack(side="right", padx=4)
-
-    win.bind('<Escape>', lambda e: (win.destroy(), 'break'))
-    win.transient(root); win.grab_set()
-    win.focus_set()
-
-__all__ = ["open_template_selector"]
+__all__ = ["SelectorView"]
